requirements_test.txt
factory_boy==2.7.0
fake-factory==0.7.2
mock==2.0.0
pytest==3.0.2
pytest-django==2.9.1
pytest-cov==2.2.1
selenium==2.53.6
tox==2.1.1
